# doc/openapi.yaml
openapi: 3.1.0
info:
  title: Homenavi External API
  version: 0.5.16
  description: |
    HTTP API exposed through the Homenavi API gateway. This document covers the
    supported public control-plane contract. Event streams are described in
    doc/asyncapi.yaml.
servers:
  - url: /
    description: Current deployment ingress origin
  - url: http://localhost
    description: Local Compose ingress
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/signup:
    post:
      tags: [Authentication]
      security: []
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignUpRequest' }
      responses:
        '201': { $ref: '#/components/responses/JSONSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/auth/login/start:
    post:
      tags: [Authentication]
      security: []
      operationId: startLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginStartRequest' }
      responses:
        '200': { $ref: '#/components/responses/JSONSuccess' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/auth/login/finish:
    post:
      tags: [Authentication]
      security: []
      operationId: finishLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginFinishRequest' }
      responses:
        '200': { $ref: '#/components/responses/JSONSuccess' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/refresh:
    post:
      tags: [Authentication]
      security: []
      operationId: refreshSession
      responses:
        '200': { $ref: '#/components/responses/JSONSuccess' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/auth/oauth/jwks.json:
    get:
      tags: [Authentication]
      security: []
      operationId: getJSONWebKeySet
      description: Public JSON Web Key Set used to verify Homenavi-issued JWTs.
      responses:
        '200':
          description: Active and overlap JWT verification keys.
          headers:
            Cache-Control:
              schema: { type: string, example: "public, max-age=300" }
          content:
            application/json:
              schema:
                type: object
                required: [keys]
                properties:
                  keys:
                    type: array
                    items:
                      type: object
                      required: [kty, use, alg, kid, n, e]
                      properties:
                        kty: { type: string, example: RSA }
                        use: { type: string, example: sig }
                        alg: { type: string, example: RS256 }
                        kid: { type: string }
                        n: { type: string }
                        e: { type: string, example: AQAB }
        '503': { description: No active signing key is available. }
  /api/auth/me:
    get:
      tags: [Authentication]
      operationId: getCurrentUser
      responses:
        '200':
          description: Current authenticated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/auth/password/reset/request:
    post:
      tags: [Authentication]
      security: []
      operationId: requestPasswordReset
      requestBody: { $ref: '#/components/requestBodies/EmailRequest' }
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
  /api/auth/password/reset/confirm:
    post:
      tags: [Authentication]
      security: []
      operationId: confirmPasswordReset
      requestBody: { $ref: '#/components/requestBodies/GenericObject' }
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
  /api/auth/password/change:
    post:
      tags: [Authentication]
      operationId: changePassword
      requestBody: { $ref: '#/components/requestBodies/GenericObject' }
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
...
  /api/auth/email/verify/confirm:
    post:
      tags: [Authentication]
      operationId: confirmEmailVerification
      requestBody: { $ref: '#/components/requestBodies/GenericObject' }
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
...
  /api/auth/2fa/verify:
    post:
      tags: [Authentication]
      operationId: verifyTwoFactor
      requestBody: { $ref: '#/components/requestBodies/GenericObject' }
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
...
  /api/auth/2fa/email/verify:
    post:
      tags: [Authentication]
      operationId: verifyEmailTwoFactor
      requestBody: { $ref: '#/components/requestBodies/GenericObject' }
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
...
  /api/auth/oauth/google/callback:
    get:
      tags: [Authentication]
      security: []
      operationId: finishGoogleOAuth
      parameters:
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        '302': { description: Redirect to the frontend after authentication }
...
  /api/auth/profile/upload-url:
    post:
      tags: [Authentication]
      operationId: createProfilePictureUploadURL
      requestBody: { $ref: '#/components/requestBodies/GenericObject' }
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
  /api/auth/profile/upload-complete:
    post:
      tags: [Authentication]
      operationId: completeProfilePictureUpload
      requestBody: { $ref: '#/components/requestBodies/GenericObject' }
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
  /api/auth/profile/upload:
    post:
      tags: [Authentication]
      operationId: uploadProfilePicture
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
...
  /api/users:
    post:
      tags: [Authentication]
      security: []
      operationId: createUserRecord
      requestBody: { $ref: '#/components/requestBodies/GenericObject' }
      responses: { '201': { $ref: '#/components/responses/JSONSuccess' } }
    get:
      tags: [Authentication]
      operationId: listUsers
      description: Backup direct user-service access. Requires the admin role.
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
...
  /api/users/{id}:
    get:
      tags: [Authentication]
      operationId: getUser
      parameters: [ { $ref: '#/components/parameters/AuthUserID' } ]
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
    patch:
      tags: [Authentication]
      operationId: updateUser
      parameters: [ { $ref: '#/components/parameters/AuthUserID' } ]
      requestBody: { $ref: '#/components/requestBodies/GenericObject' }
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
    delete:
      tags: [Authentication]
      operationId: deleteUser
      parameters: [ { $ref: '#/components/parameters/AuthUserID' } ]
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
  /api/users/{id}/lockout:
    post:
      tags: [Authentication]
      operationId: lockOutUser
      parameters: [ { $ref: '#/components/parameters/AuthUserID' } ]
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
...
  /api/auth/users/{id}:
    get:
      tags: [Authentication]
      operationId: getAuthUser
      parameters: [ { $ref: '#/components/parameters/AuthUserID' } ]
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
    patch:
      tags: [Authentication]
      operationId: updateAuthUser
      parameters: [ { $ref: '#/components/parameters/AuthUserID' } ]
      requestBody: { $ref: '#/components/requestBodies/GenericObject' }
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
  /api/auth/users/{id}/lockout:
    post:
      tags: [Authentication]
      operationId: lockOutAuthUser
      parameters: [ { $ref: '#/components/parameters/AuthUserID' } ]
      responses: { '200': { $ref: '#/components/responses/JSONSuccess' } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT sent as an Authorization bearer token. Browser clients may also use the auth cookie.
  parameters:
...
    AuthUserID: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
  requestBodies:
    EmailRequest:
      required: true
      content: { application/json: { schema: { type: object, required: [email], properties: { email: { type: string, format: email } } } } }
    GenericObject:
      required: true
      content: { application/json: { schema: { type: object, additionalProperties: true } } }
  responses:
    JSONSuccess:
      description: Successful JSON response
      content: { application/json: { schema: { type: object, additionalProperties: true } } }
...
    BadRequest: { description: Request validation failed }
    Unauthorized: { description: Authentication is required or invalid }
...
  schemas:
    SignUpRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
        name: { type: string }
    LoginStartRequest:
      type: object
      required: [email]
      properties: { email: { type: string, format: email } }
    LoginFinishRequest:
      type: object
      required: [email, code]
      properties:
        email: { type: string, format: email }
        code: { type: string }
    User:
      type: object
      required: [id, email]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        role: { type: string, enum: [admin, resident] }
...